requirements.txt
Pillow # gerenciar imagens
google-generativeai
python-telegram-bot
firebase-admin==6.6.0 
# dot_env
python-dotenv

# Formating code:
flake8
black
isort